portinfo: Use benign RPCs to discover what ports are all about
Most servers have at least one RPC which is benign enough to be used to
identify whether a port supports the server and provides some useful
information to identify what the port is about. This adds support for
most of these, allowing to identify auth ports, file descriptors,
sockets, etc.