Fix parallel builds
As explained on
https://www.gnu.org/software/automake/manual/html_node/Multiple-Outputs.html

a b: c
	foo -o a -o b

is actually equivalent to

a: c
	foo -o a -o b
b: c
	foo -o a -o b

and can thus break in parallel builds, since the foo command may then be
called twice. Also, we are here using pattern rules. The automake manual
contains various levels of fixing this, using a mere stamp file should
be enough for us.