Stop checking if MiG supports retcode.
Due to MiG requiring the subsystem to be defined early in order to know the
size of a port, this was causing a division by zero error during ./configure.
We could have just move subsystem to the top of the snippet, however it is
simpler to just remove the check given that we have no plans to use some other
MiG anyway.

We force RETURN_CODE_ARG to use retcode which is a no-op.
Message-Id: <ZFhs3KSeGhuubS8T@jupiter.tail36e24.ts.net>