lwip: Allocate the loopback netif by default
The translator received a null `netif_list` during initialization, this
caused a few bugs.

When started without parameters, the translator didn't add any new
interface to `netif_list`, and that broke any subsequent fsysopts over
the translator, as the stack was being initialized again instead of
being reconfigured.

DHCP was broken because the translator is usually installed without
parameters, which are supposed to be added by the DHCP client through
fsysopts.

The absence of an allocated `netif_list` also prevented configuring a
loopback interface.

After these changes, starting the translator always allocates one
interface and configures it as loopback.
Message-ID: <20231202101401.11542-2-jlledom@mailfence.com>