Restructure argument setup in hashbang
We do a few things here:
- Move search_path to the scope where it is used to make dependencies
  more clear.
- Have a separate variable to store the file name we eventually need to
  free and move the free logic to happen in a single place.

Both of this allows us to still free the name even if a fault is generated and
also avoids a compiler warning as we try to assign a 'const char*'
file_name_exec to a 'char *', making it more clear to what exactly we
need to free. I also believe 'error' in line 245 was not initialized in
case 'file_name_exec' is used and this fixes that too.
Message-ID: <20240121210757.1900938-1-flaviocruz@gmail.com>