rumpdisk: do not open device if block size is 0
Currently, if we do:

$ ls /dev/cd0/

The computer seems to get stuck, caused by the divide by 0 in the
rumpdisk server in device_get_status. I noticed that if we have no disk in the
cdrom device, we can still open it but block and media size will be 0
and the message "cd0 dos partition I/O error" will be printed to the
console. To avoid this problem, we check the block size and throw an error
when it is 0. This also works correctly when a disk actually exists.

This should help fix the perl and likely the vim test suites that are
currently failing in https://buildd.debian.org/.
Message-ID: <Zd_8XjcHcbNIp5NM@mars.tail36e24.ts.net>