MAKEDEV: when creating devices, ensure the underlying files are either block/char devices or directories
The perl test suite has a test where it reads all the block or char devices
under /dev without following the translators. Then it compares it against a
list of devices that read the translated nodes stat info.

The patch changes how the the device files are created initially so that the stat information
is identical and makes the Hurd environment appear more similar to other operating
systems.
Message-ID: <Zef__LrgPLXp9WHG@jupiter.tail36e24.ts.net>